fix(onboard): align Gemini credential guidance
